refactor IA and model directory in Game directory
