+++ Game/Model/Knight.py
